feat: 상품 주문 facade 구현

+++ apps/commerce-api/src/main/java/com/loopers/application/point/PointResult.java
@@ -3,7 +3,7 @@ package com.loopers.application.point;
 import com.loopers.domain.point.Point;
 import com.loopers.domain.point.PointInfo;
 
-public record PointResult(Long userId, int balance) {
+public record PointResult(Long userId, long balance) {
     public static PointResult from(PointInfo info) {
         return new PointResult(
                 info.userId(),
