+++ apps/commerce-api/src/main/java/com/loopers/domain/point/PointCommand.java
@@ -3,9 +3,9 @@ package com.loopers.domain.point;
 public class PointCommand {
     public record Charge(
             Long userId,
-            int amount
+            long amount
     ) {
-        public static Charge of(Long userId, int amount) {
+        public static Charge of(Long userId, long amount) {
             return new Charge(userId, amount);
         }
     }
@@ -13,9 +13,9 @@ public class PointCommand {
     public record Using(
             Long userId,
             Long orderId,
-            int amount
+            long amount
     ) {
-        public static Using of(Long userId, long orderId,int amount) {
+        public static Using of(Long userId, long orderId,long amount) {
             return new Using(userId, orderId, amount);
         }
     }
